// Cyclecount_TA_FRONT/src/app/Model/LICC.model.ts
export interface LICCModel {
  id: any,
  nidat:string,
  ttext:string,
  salk3: number,
  ageing_days:number,

  mandt: number
  abcin: string
  lgnum: number
  ivnum: string
  ivpos: string
  lqnum: string
  nanum: string
  istat: string
  nvers: string
  iseit: string
  lgtyp: string
  lgpla: string
  plpos: string
  matnr: string
  werks: string
  charg: string
  letyp: string
  anzle: number
  menge: number
  meins: string
  menga: number
  altme: string
  umrez: number
  umren: number
  bestq: string
  sobkz: string
  sonum: string
  wdatu: string
  gesme: number
  idatu: string
  kzinv: string
  irnum: string
  tanum: string
  tapos: string
  lenum: string
  vfdat: string
  lgort: string
  uname: string
  quinv: string
  idqua: string
  dbgez: string
  key: number
  lvorm: string
  maktx: string
  kreuz: string
  pagno: number
  linno: number
  bemer: string
  abwei: number
  vzchn: string
  ausbu: string
  bwkey: string
  bukrs: string
  waers: string
  vprsv: string
  verpr: number
  stprs: number
  peinh: number
  cwert: string
  dwert: number
  noval: string
  lsonr: string
  cntstatus: string
  statustype: string
  idusin: any
  idusup: any
  plantID: number
  fromLICC: boolean
  ymd8in: string
  ymd8up: string
  insertedBy: any
  counter: any
  plant: any
}